+++ src/components/BuildInfo/BuildInfo.stories.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import BuildInfo from "./BuildInfo";
+
+storiesOf("BuildInfo", module)
+  .add("default", () => <BuildInfo />)
+  .add("temporary install", () => <BuildInfo temporaryInstall={true} />);
