+++ src/components/Button/Button.stories.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+import { action } from "@storybook/addon-actions";
+
+import Button from "./Button";
+
+storiesOf("Button", module).add("default", () => (
+  <Button onClick={action("clicked")}>Click me!</Button>
+));
