+++ src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-import Alert from "./Alert";
+import Alert from "../Alert";
 
 type State = {
   hasError: boolean;
