+++ src/components/Loader/Loader.stories.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import Loader from "./Loader";
+
+storiesOf("Loader", module).add("default", () => <Loader />);
