+++ src/components/PopupContainer/PopupContainer.stories.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import PopupContainer from "./PopupContainer";
+
+storiesOf("PopupContainer", module).add("default", () => <PopupContainer />);
